Rework how multiple return nodes are handled
The new version makes more sense for how nodes are generally laid out on the tree.

If there's a command and a group as sibling nodes, return those as an array, and let the implementation handle detecting that. This allows for paramaterized overloads to be displayed to users for executable groups.

The previous implementation of the multi-node method also just didn't work, and this was evident even before I got around to writing tests for it.

The code should be more logical and readable now, as well.

The old code in the help service was also changed to a switch expression for clarity, taking it down to all of 5 lines.